# fix: parallel async datachange_notification calls