### Description:
Slightly improve performance when `datachange_notification` handler method is a coroutine, by using `asyncio.gather` to ensure calls are parallelised 

#### Test Environment:
MacBook Pro (16-inch, 2019)
Python 3.8.10
opcua-asyncio[master]

#### Data:
Subscription for 1000 variables changing with frequency of 1Hz
Subscription publish interval - 5 sec

#### Handler:

```py
class OpcUaDataGatewayService(BaseOPCService):
    ...

     async def process_event(self, event, node, value):
        event_data = await self.data_provider.serialize(event, value=value)  # get event node attributes as dict
        measurement = self._make_measurement(node_measurement_config, **event_data)
        await self.send_message(measurement)  # buffering happens in this function so it's pretty fast

    async def datachange_notification(self, node, value, event):
        await self.process_event(event, node, value)
```

#### Results:
Timings before:
> 2022-04-15 13:05:05,426, [root] INFO: "Completed _call_datachange. Elapsed time: 0.3814258575439453 seconds"
> 2022-04-15 13:05:10,533, [root] INFO: "Completed _call_datachange. Elapsed time: 0.402393102645874 seconds"
> 2022-04-15 13:05:15,444, [root] INFO: "Completed _call_datachange. Elapsed time: 0.3754401206970215 seconds"
> 2022-04-15 13:05:20,477, [root] INFO: "Completed _call_datachange. Elapsed time: 0.3901801109313965 seconds"
> 2022-04-15 13:05:25,485, [root] INFO: "Completed _call_datachange. Elapsed time: 0.3795757293701172 seconds"

Timings after fix is applied:
> 2022-04-15 13:11:22,351, [root] INFO: "Completed _call_datachange. Elapsed time: 0.37169408798217773 seconds"
> 2022-04-15 13:11:27,418, [root] INFO: "Completed _call_datachange. Elapsed time: 0.36106228828430176 seconds"
> 2022-04-15 13:11:32,453, [root] INFO: "Completed _call_datachange. Elapsed time: 0.37993907928466797 seconds"
> 2022-04-15 13:11:37,420, [root] INFO: "Completed _call_datachange. Elapsed time: 0.3285830020904541 seconds"
> 2022-04-15 13:11:42,278, [root] INFO: "Completed _call_datachange. Elapsed time: 0.35353779792785645 seconds"