@oroulet 2022-04-15
Good idea